added basic admins page

+++ actions/deleteCategory.php
@@ -1,7 +1,12 @@
 <?php
     if(is_logged_in() && is_admin() && isset($_POST['action']) && $_POST['action'] == 'delete') {
-        $id = $_POST['id'];
-        $sql = "delete from categories where id = ?";
-        $result = $connection->execute_query($sql, array($id));
+        if($_SESSION['delete_downloads'] == 1) {
+            $id = $_POST['id'];
+            $sql = "delete from categories where id = ?";
+            $result = $connection->execute_query($sql, array($id));
+
+        } else {
+            $error = "У вас недостаточно прав";
+        }
     }
 ?>
